requirements.txt

typer
pyyaml
numpy
pandas
scikit-learn
lightgbm
xgboost
catboost
joblib
scipy
optuna
kaggle
torch
torchvision
timm
pillow
transformers
accelerate
sentencepiece
